feat: package timeout for k8s communication (#158)
* feat: package timeout for k8s communication


* feat: const instead of var


* feat: move utils to utils.go


---------
